addresses #26 install isseus and #12 live-reload
