+++ index.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var app = require('./app');
 var url = require("url");
-
+var exec = require( "child_process" ).exec;
 
 app.set('port', (process.env.PORT || 5000));
 
@@ -20,3 +20,14 @@ require('./services/users');
 app.listen(app.get('port'), function() {
   console.log('Node app is running on port', app.get('port'));
 });
+
+
+if ( process.argv.indexOf( "--develop" ) !== -1 ) {
+  //is dev mode so do live reload
+  var child = exec( "node_modules/.bin/steal-tools live-reload", {
+    cwd: process.cwd() + "/public"
+  });
+
+  child.stdout.pipe( process.stdout );
+  child.stderr.pipe( process.stderr );
+}
