+++ post_install.js
@@ -1,0 +1,11 @@
+var fs = require("fs");
+fs.unlinkSync(__dirname+"/node_modules/.bin/npm");
+
+var exec = require( "child_process" ).exec;
+
+var child = exec( "npm install", {
+	cwd: process.cwd() + "/public"
+});
+
+child.stdout.pipe( process.stdout );
+child.stderr.pipe( process.stderr );
